# src/main/java/com/possable/controller/PrinterController.java
package com.possable.controller;

import com.possable.service.PrinterService;
import jakarta.validation.Valid;
import jakarta.validation.constraints.NotBlank;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.Map;

@RestController
@RequestMapping("/printers")
public class PrinterController {

    private final PrinterService printerService;

    public PrinterController(PrinterService printerService) {
        this.printerService = printerService;
    }

    @GetMapping
    public ResponseEntity<List<PrinterService.Printer>> listPrinters() {
        return ResponseEntity.ok(printerService.listPrinters());
    }

    public record RegisterPrinterRequest(@NotBlank String name, @NotBlank String category, String description) {}

    @PostMapping
    public ResponseEntity<Map<String, Object>> registerPrinter(@Valid @RequestBody RegisterPrinterRequest req) {
        var p = printerService.registerPrinter(req.name(), req.category(), req.description());
        return ResponseEntity.status(201).body(Map.of("id", p.id(), "name", p.name(), "category", p.category()));
    }
} 